Code refactor - Move AWS profile management to it's own class. Modularize prompt generation classes. Minor updates to loggers
